Move schema to separated file
